Tried to re-adjust CommandServiceMessage architecture (see body);
Fixed branching issue (accidentally committed to the commandService
branch) & refactored GenericProg into an abstract DevdasCore (and moved
concrete commands in a new class, DevdasCoreTester). Partially removed
threading functionality in CommandProcessors (will implement later).

Readjusted CommandServiceMessage to allow for flexible use of
parameters, but there is a parsing issue that must be fixed. The new
param field, which is a Map, cannot be conveniently parsed when received
as a String through the Subscriber (this is due to an issue found in the
read(JSONObject) method in the CommandServiceMessage class).