Found solution to bug, but ramifications need to be discussed;
Made the mapping for AgentReactions (in DevdasCore) expandable so that
an interest can have multiple reactions. Each key-value pair is now a
String and a ArrayList of AgentReactions (instead of a singular
AgentReaction).

This solution makes the code more complex and hard to implement, in that
it is harder to access a lone AgentReaction within the ArrayList without
prior knowledge to the structure of the list (methods must be provided
to make it easier to implement), but it provides a complete solution to
the aforementioned bug.